[AIRFLOW-3888] HA for Hive metastore connection (#4708)
Creating a connection to a metastore with two hosts for high availability (eg connection 1, connection 2) is not possible because the entire value entered is taken. For our needs, it is necessary to go through subsequent hosts and connect to the first working.

This change allows you to check and then connect to a working metastore.

(cherry picked from commit 355bd56282e6a684c5c060953e9948ba2260aa37)